refactor and add playground test
