+++ impeller/display_list/BUILD.gn
@@ -12,6 +12,8 @@ impeller_component("display_list") {
     "display_list_image_impeller.h",
     "nine_patch_converter.cc",
     "nine_patch_converter.h",
+    "vertices_converter.cc",
+    "vertices_converter.h",
   ]
 
   public_deps = [
